Update compose layer
